Split types and response models
